fix: align TPC-H Parquet row-group validation with TPC-DS (#424)
* fix: align TPC-H Parquet row-group validation with TPC-DS

Reject non-positive TPC-H row-group byte sizes during clap parsing, matching the TPC-DS behavior introduced in #349. Clarify the option help for both benchmarks and cover the parsing boundaries.


* style: format row-group validation assertion


* docs: simplify row-group option help


* refactor: use Clap range validation for TPC-DS row groups

Replace the custom usize parser with Clap's ranged parser and update the rejection assertion for its standard diagnostics.


* docs: keep row-group option help concise

Leave the positive-value constraint to Clap's argument validation diagnostics.


* refactor: share row-group parsing with clear CLI errors

Use a signed intermediate to report must be greater than zero for both CLI types, then check conversion to the original i64 or usize range.


* refactor: standardize row-group byte sizes on i64

Use i64 throughout both CLI paths and TPC-DS planning, removing the unchecked usize cast and simplifying the shared parser. Both commands now accept the same positive i64 range.


* test: remove redundant maximum row-group planning case


* test: consolidate row-group boundary coverage

Keep the minimum-value assertion in the shared-parser test and remove the separate TPC-H CLI test.


* test: leave existing TPC-DS rejection assertion unchanged


---------

+++ tpcgen-cli/src/tpcds_cli/parquet.rs
@@ -100,7 +100,7 @@ fn column_encodings_for_table(
 pub(super) struct Parquet {
     output_dir: PathBuf,
     compression: Compression,
-    row_group_bytes: usize,
+    row_group_bytes: i64,
     num_threads: usize,
     column_encodings: Option<Vec<(String, Encoding)>>,
 }
@@ -109,7 +109,7 @@ impl Parquet {
     pub(super) fn new(
         output_dir: PathBuf,
         compression: Compression,
-        row_group_bytes: usize,
+        row_group_bytes: i64,
         num_threads: usize,
         column_encodings: Option<Vec<(String, Encoding)>>,
     ) -> Self {
